refactor(tracklet): share scale-aware noise machinery between BoT-SORT and McByte
BoTSORTTracklet and McByteTracklet held byte-identical copies of the
scale-aware Kalman-noise code: _SIGMA_* constants, _configure_initial_noise,
_build_process_noise, _build_measurement_noise, _set_scale_aware_noise,
_current_wh, both _refresh_*_noise_from_state, the _clamp_* helpers, and
(from #578) a duplicated _diagonal_matrix helper.

- _diagonal_matrix moves to trackers.utils.matrix.
- The noise machinery moves to a new ScaleAwareNoiseTracklet(BaseTracklet)
  in trackers.utils.scale_aware_tracklet; both tracklets now subclass it
  and keep only their own __init__/update/predict/get_state_bbox/apply_cmc.
- The duplicated _diagonal_matrix regression tests collapse into one
  tests/core/test_matrix.py.

Neither class is exported from trackers/__init__.py, so the public API is
unchanged. No behavior change: method bodies are moved verbatim and the
full non-integration suite passes (1560 passed).
