+++ src/trackers/core/mcbyte/tracklet.py
@@ -8,32 +8,13 @@ from __future__ import annotations
 
 import numpy as np
 
-from trackers.utils.base_tracklet import BaseTracklet
 from trackers.utils.cmc import CMC
-from trackers.utils.converters import xyxy_to_xywh
 from trackers.utils.predict_timing import FIXED_RATE_TIMING, PredictTiming
-from trackers.utils.state_representations import (
-    BaseStateEstimator,
-    XCYCSRStateEstimator,
-    XCYCWHStateEstimator,
-    XYXYStateEstimator,
-)
+from trackers.utils.scale_aware_tracklet import ScaleAwareNoiseTracklet
+from trackers.utils.state_representations import BaseStateEstimator, XCYCWHStateEstimator
 
 
-def _diagonal_matrix(values: list[float]) -> np.ndarray:
-    """Construct a diagonal matrix directly from known values.
-
-    Construct-only: always returns float64, regardless of the input
-    values' type. Unlike ``np.diag``, does not extract a diagonal from
-    a 2-D input.
-    """
-    size = len(values)
-    matrix = np.zeros((size, size), dtype=np.float64)
-    matrix.flat[:: size + 1] = values
-    return matrix
-
-
-class McByteTracklet(BaseTracklet):
+class McByteTracklet(ScaleAwareNoiseTracklet):
     """Tracklet for the McByte tracker.
 
     Uses ``XCYCWHStateEstimator`` (center + width/height) by default,
@@ -41,7 +22,8 @@ class McByteTracklet(BaseTracklet):
 
     * **Scale-aware noise**: ``Q``, ``R`` and the initial ``P`` are computed
       from the current width / height of the tracked object each frame, so
-      that uncertainty scales with object size.
+      that uncertainty scales with object size (see
+      :class:`~trackers.utils.scale_aware_tracklet.ScaleAwareNoiseTracklet`).
     * **Width / height clamping** after every predict and update step.
     * ``predict()`` increments ``time_since_update``: unmatched tracks are
       never explicitly fed ``update(None)``.
@@ -50,11 +32,6 @@ class McByteTracklet(BaseTracklet):
     * ``apply_cmc(H)`` applies a 2x3 affine camera-motion transform to the
       internal Kalman state and covariance.
     """
-
-    # Noise sigma constants (scale-aware noise for McByte)
-    _SIGMA_P: float = 0.05
-    _SIGMA_V: float = 0.00625
-    _SIGMA_M: float = 0.05
 
     def __init__(
         self,
@@ -66,155 +43,6 @@ class McByteTracklet(BaseTracklet):
         # Count initial bbox as first successful update so that
         # number_of_successful_updates starts at 1.
         self.number_of_successful_updates = 1
-
-    def _configure_initial_noise(self, bbox: np.ndarray) -> None:
-        """Set initial P, Q, R based on the first detection's size."""
-        measurement = xyxy_to_xywh(bbox)
-        w, h = float(measurement[2]), float(measurement[3])
-        self._set_scale_aware_noise(w, h)
-
-    def _build_process_noise(self, w: float, h: float) -> np.ndarray:
-        """Build the scale-aware process-noise covariance (Q) for the given box size."""
-        sp, sv = self._SIGMA_P, self._SIGMA_V
-        if isinstance(self.state_estimator, XCYCSRStateEstimator):
-            s = np.sqrt(max(w * h, 1e-6))
-            return _diagonal_matrix(
-                [
-                    (sp * w) ** 2,
-                    (sp * h) ** 2,
-                    (sp * s) ** 2,
-                    (sp * 1.0) ** 2,
-                    (sv * w) ** 2,
-                    (sv * h) ** 2,
-                    (sv * s) ** 2,
-                ]
-            )
-        return _diagonal_matrix(
-            [
-                (sp * w) ** 2,
-                (sp * h) ** 2,
-                (sp * w) ** 2,
-                (sp * h) ** 2,
-                (sv * w) ** 2,
-                (sv * h) ** 2,
-                (sv * w) ** 2,
-                (sv * h) ** 2,
-            ]
-        )
-
-    def _build_measurement_noise(self, w: float, h: float) -> np.ndarray:
-        """Build the scale-aware measurement-noise covariance (R) for the given box size."""
-        sm = self._SIGMA_M
-        if isinstance(self.state_estimator, XCYCSRStateEstimator):
-            s = np.sqrt(max(w * h, 1e-6))
-            return _diagonal_matrix([(sm * w) ** 2, (sm * h) ** 2, (sm * s) ** 2, (sm * 1.0) ** 2])
-        return _diagonal_matrix([(sm * w) ** 2, (sm * h) ** 2, (sm * w) ** 2, (sm * h) ** 2])
-
-    def _set_scale_aware_noise(self, w: float, h: float) -> None:
-        """Set the initial Q, R and P from the first detection's size."""
-        sp, sv = self._SIGMA_P, self._SIGMA_V
-        Q = self._build_process_noise(w, h)
-        R = self._build_measurement_noise(w, h)
-
-        if isinstance(self.state_estimator, XCYCSRStateEstimator):
-            s = np.sqrt(max(w * h, 1e-6))
-            state_covariance = _diagonal_matrix(
-                [
-                    (2 * sp * w) ** 2,
-                    (2 * sp * h) ** 2,
-                    (2 * sp * s) ** 2,
-                    (2 * sp * 1.0) ** 2,
-                    (10 * sv * w) ** 2,
-                    (10 * sv * h) ** 2,
-                    (10 * sv * s) ** 2,
-                ]
-            )
-        else:
-            state_covariance = _diagonal_matrix(
-                [
-                    (2 * sp * w) ** 2,
-                    (2 * sp * h) ** 2,
-                    (2 * sp * w) ** 2,
-                    (2 * sp * h) ** 2,
-                    (10 * sv * w) ** 2,
-                    (10 * sv * h) ** 2,
-                    (10 * sv * w) ** 2,
-                    (10 * sv * h) ** 2,
-                ]
-            )
-        self.state_estimator.set_kf_covariances(measurement_noise=R, process_noise=Q, state_covariance=state_covariance)
-
-    def _current_wh(self) -> tuple[float, float]:
-        """Return the current box width/height, clamped away from zero.
-
-        The ``XCYCWHStateEstimator`` path reconstructs each pair of corners before subtracting them. Reading the stored
-        width and height directly would skip the center-dependent rounding performed by ``state_to_bbox()``. The fast
-        path applies only when the estimator is exactly ``XCYCWHStateEstimator`` and its state has dtype ``float64``.
-        Subclasses and states with other dtypes fall back to ``state_to_bbox()`` so overridden conversion behavior and
-        the decoder's float64 corner arithmetic is preserved.
-        """
-        state = self.state_estimator.kf.state
-        if type(self.state_estimator) is XCYCWHStateEstimator and state.dtype == np.float64:
-            x_center = state[0, 0]
-            y_center = state[1, 0]
-            half_w = state[2, 0] * 0.5
-            half_h = state[3, 0] * 0.5
-            w = max(float((x_center + half_w) - (x_center - half_w)), 1e-3)
-            h = max(float((y_center + half_h) - (y_center - half_h)), 1e-3)
-            return w, h
-
-        bbox = self.state_estimator.state_to_bbox()
-        w = max(float(bbox[2] - bbox[0]), 1e-3)
-        h = max(float(bbox[3] - bbox[1]), 1e-3)
-        return w, h
-
-    def _refresh_process_noise_from_state(self) -> None:
-        """Recompute Q (process noise) from the current bbox size.
-
-        Only ``predict()`` reads ``process_noise`` (see ``KalmanFilter.predict``), so this must not be called from
-        ``update()`` — it would be overwritten by the next ``predict()`` before ever being read.
-        """
-        w, h = self._current_wh()
-        self.state_estimator.set_kf_covariances(process_noise=self._build_process_noise(w, h))
-
-    def _refresh_measurement_noise_from_state(self) -> None:
-        """Recompute R (measurement noise) from the current bbox size.
-
-        Only ``update()`` reads ``measurement_noise`` (see ``KalmanFilter.update``), so this must not be called from
-        ``predict()`` — it would be overwritten by the next ``update()`` before ever being read.
-        """
-        w, h = self._current_wh()
-        self.state_estimator.set_kf_covariances(measurement_noise=self._build_measurement_noise(w, h))
-
-    @staticmethod
-    def _clamp_xyxy_state(kf_x: np.ndarray) -> None:
-        """Ensure XYXY state keeps valid box corners."""
-        if kf_x[2, 0] <= kf_x[0, 0]:
-            kf_x[2, 0] = kf_x[0, 0] + 1e-3
-        if kf_x[3, 0] <= kf_x[1, 0]:
-            kf_x[3, 0] = kf_x[1, 0] + 1e-3
-
-    @staticmethod
-    def _clamp_xcycwh_state(kf_x: np.ndarray) -> None:
-        """Ensure XCYCWH state keeps positive width and height."""
-        kf_x[2, 0] = max(kf_x[2, 0], 1e-3)
-        kf_x[3, 0] = max(kf_x[3, 0], 1e-3)
-
-    @staticmethod
-    def _clamp_xcycsr_state(kf_x: np.ndarray) -> None:
-        """Ensure XCYCSR state keeps positive scale and aspect ratio."""
-        kf_x[2, 0] = max(kf_x[2, 0], 1e-3)
-        kf_x[3, 0] = max(kf_x[3, 0], 1e-3)
-
-    def _clamp_state_bbox(self) -> None:
-        """Clamp geometric components based on active state representation."""
-        kf_x = self.state_estimator.kf.state
-        if isinstance(self.state_estimator, XYXYStateEstimator):
-            self._clamp_xyxy_state(kf_x)
-        elif isinstance(self.state_estimator, XCYCWHStateEstimator):
-            self._clamp_xcycwh_state(kf_x)
-        elif isinstance(self.state_estimator, XCYCSRStateEstimator):
-            self._clamp_xcycsr_state(kf_x)
 
     def update(self, bbox: np.ndarray) -> None:
         """Update tracklet with a new observation.
